backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
openai>=1.60.0
python-multipart>=0.0.20
pydantic>=2.10.0
reportlab>=4.2.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
